+++ kiosc/users/admin.py
@@ -13,7 +13,7 @@ class MyUserChangeForm(UserChangeForm):
 class MyUserCreationForm(UserCreationForm):
 
     error_message = UserCreationForm.error_messages.update(
-        {'duplicate_username': 'This username has already been taken.'}
+        {"duplicate_username": "This username has already been taken."}
     )
 
     class Meta(UserCreationForm.Meta):
@@ -25,7 +25,7 @@ class MyUserCreationForm(UserCreationForm):
             User.objects.get(username=username)
         except User.DoesNotExist:
             return username
-        raise forms.ValidationError(self.error_messages['duplicate_username'])
+        raise forms.ValidationError(self.error_messages["duplicate_username"])
 
 
 @admin.register(User)
@@ -33,7 +33,7 @@ class MyUserAdmin(AuthUserAdmin):
     form = MyUserChangeForm
     add_form = MyUserCreationForm
     fieldsets = (
-        ('User Profile', {'fields': ('name',)}),
+        ("User Profile", {"fields": ("name",)}),
     ) + AuthUserAdmin.fieldsets
-    list_display = ('username', 'name', 'is_superuser')
-    search_fields = ['name']
+    list_display = ("username", "name", "is_superuser")
+    search_fields = ["name"]