+++ kiosc/users/tests/test_models.py
@@ -7,5 +7,5 @@ class TestUser(TestCase):
 
     def test__str__(self):
         self.assertEqual(
-            self.user.__str__(), 'testuser'
+            self.user.__str__(), "testuser"
         )  # This is the default username for self.make_user()